crypto_macro_alert_bot_v6/requirements.txt
python-dotenv 
python-telegram-bot==20.7
python-binance 
pandas
ta
matplotlib
mplfinance
requests
numpy